Silence `gcc -Wall'.
* buffer.c (buf_initialize): Remove unnecessary typecasts by using
size_t instead of int or bool as args.
* buffer.c, import.c, log-buffer.c, ms-buffer.c, zlib.c: Change all
callers.
* buffer.h, cvs.h: Update protos.
* client.c, ls.c, main.c, rcs.c, root.c, server.c: Remove unused vars,
add parens, as requested by `gcc -Wall'.

CVS-ID: src/ChangeLog 1.2875
CVS-ID: src/buffer.c 1.41
CVS-ID: src/buffer.h 1.18
CVS-ID: src/client.c 1.382
CVS-ID: src/cvs.h 1.301
CVS-ID: src/import.c 1.152
CVS-ID: src/log-buffer.c 1.13
CVS-ID: src/ls.c 1.15
CVS-ID: src/main.c 1.217
CVS-ID: src/ms-buffer.c 1.3
CVS-ID: src/rcs.c 1.315
CVS-ID: src/root.c 1.84
CVS-ID: src/server.c 1.373
CVS-ID: src/zlib.c 1.22

+++ src/cvs.h
@@ -562,7 +562,7 @@ time_t get_date (char *date, struct timeb *now);
 int Create_Admin (const char *dir, const char *update_dir,
                   const char *repository, const char *tag, const char *date,
                   int nonbranch, int warn, int dotemplate);
-int expand_at_signs (const char *, off_t, FILE *);
+int expand_at_signs (const char *, size_t, FILE *);
 
 /* Locking subsystem (implemented in lock.c).  */
